fix: resolve 3 bugs from testing
Bug #1 (CRITICAL): Fix SQL date comparison in BuyerSelectionService
- Cast date properly: session_date = ($1::date - INTERVAL '1 day')::date
- Fixes 'operator does not exist: date = interval' error

Bug #2 (CRITICAL): Fix array type mismatch in OrderController.getHistory
- Cast userId to integer in array comparison: ARRAY[$1::integer]
- Fixes 'operator does not exist: integer[] @> text[]' error

Bug #3 (MINOR): Update test account password hashes
- Re-generate bcrypt hashes for Admin123! and User123!
- Admin: admin@lunchfund.com (Admin123!)
- Users: user1-5@test.com (User123!)

All test accounts now work with correct passwords.

+++ backend/src/controllers/OrderController.js
@@ -287,7 +287,7 @@ class OrderController {
           ls.total_bill,
           ls.amount_per_person,
           lo.created_at as joined_at,
-          CASE WHEN ls.buyer_ids @> ARRAY[$1] THEN true ELSE false END as was_buyer,
+          CASE WHEN ls.buyer_ids @> ARRAY[$1::integer] THEN true ELSE false END as was_buyer,
           CASE WHEN ls.payer_id = $1 THEN true ELSE false END as was_payer
         FROM lunch_orders lo
         JOIN lunch_sessions ls ON lo.session_id = ls.id
